Placement stats

+++ src/main/java/no/vebb/f1/controller/admin/FlagController.java
@@ -2,6 +2,8 @@ package no.vebb.f1.controller.admin;
 
 import java.util.List;
 
+import no.vebb.f1.util.domainPrimitive.Year;
+import no.vebb.f1.util.exception.YearFinishedException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
@@ -56,6 +58,10 @@ public class FlagController {
             @RequestParam("sessionType") String sessionType) {
         try {
             RaceId validRaceId = new RaceId(raceId, db);
+            Year year = db.getYearFromRaceId(validRaceId);
+            if (db.isFinishedYear(year)) {
+                throw new YearFinishedException("Year '" + year + "' is over and the flags can't be changed");
+            }
             Flag validFlag = new Flag(flag, db);
             SessionType validSessionType = new SessionType(sessionType, db);
             if (!isValidRound(round)) {
@@ -72,6 +78,10 @@ public class FlagController {
     @PostMapping("/delete")
     @Transactional
     public ResponseEntity<?> deleteFlag(@RequestParam("id") int id) {
+        Year year = db.getYearFromFlagId(id);
+        if (db.isFinishedYear(year)) {
+            throw new YearFinishedException("Year '" + year + "' is over and the flags can't be changed");
+        }
         db.deleteFlagStatsById(id);
         return new ResponseEntity<>(HttpStatus.OK);
     }
